+++ solutions/ng-rx/src/app/services/todo.service.ts
@@ -1,4 +1,4 @@
-import { Todo } from './../models/todo.model';
+import { Todo } from './../store/todo.model';
 import { Injectable } from '@angular/core';
 import { Http } from '@angular/http';
 import { Observable } from 'rxjs/Observable';
