+++ solutions/ng-rx/src/app/store/todo.model.ts
